Remove unneeded function with cryptic name
There was a cryptic function duplicates() in the pdt_cad_module.py that
was called from pdt_xall.py - a legacy from the code borrowed from
TinyCAD Add-on. This seems to be a superfluous function that does not
need to exist, so therefore has been removed.

This commit addresses the following Blender review notes:

- "This function seems quite cryptic for what it actually does (...)"