Add script for cutting official Blender PDT releases
Every time we need to sync the latest release to the Blender addons/ repo,
running this script will create a folder with the correct name and files.

This folder can then be copied into the Blender addons/ repo where a commit
can be made with a suitable commit message outlining the changes since the
last sync from the PDT github master branch.