Fix: Small Error in Absolute based Functions
This fixes a small obscure error in Absolute mode whereby Bmesh was not always found
if Absolute mode was used. This became apparent if no object was selected and an
attempt was made to add a new vertex, or extrude vertices. Other checks did not trap
for this error. These traps originally were stored in the individual functions, now all traps are
 moved to the command parse function.