Fix: Issue with Parts Library File
An issue occurred if you opened Blender with a new file. The library file pointed relative to
the Blender App, not the working directory so was not found. Parts Library file is now moved
to the Parts Library Menu, so must be set for a new blend file, existing saved blend files are not
affected by this change.

Error message altered to reflect the option of having this Parts Library anywhere.