Enhancement: View Normal Operation
This enhancement enables the user to work always Normal to the View
as defined by "Working Plane" in PDT Design Functions Section.

For example: If the working plane is set to Front and the Operation is "Extrude Geometry"
the system will extrude along the Y axis by the amount specified in the "Distance" Entry Box.
Specifying Working Plane as "View" cause the system to work always Normal to the current
view orientation, i.e. into, or out of your screen.

This enhancement is only available for Blender 2.9x builds.