Bug Fix - Fillet Method - T95442
Some versions of Linux and Windows show an error with the value '_segments' in pdt_command.py
(the Fillet method).

This error manifests as not recognising the number of segments as a float value as previously.

This error does not manifest itself on MacOS, or Ubuntu versions of Blender. The values extracted
are strings of the form "4.0" so for Python to get this into an integer format it is necessary to use
a double function 'int(float([Value String]))' in line 1047 of pdt_command.py.

This modification has been checked up to Blender version 3.2 Alpha.